Refactored WaiverBlock => InDemandBlock, simplified component logic, implemented on trainings
