+++ frontend/src/components/__tests__/InDemandBlock.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom/extend-expect";
-import { WaiverBlock } from "../WaiverBlock";
+import { InDemandBlock } from "../InDemandBlock";
 import { useMediaQuery } from "@material-ui/core";
 
 jest.mock("@material-ui/core", () => ({
@@ -13,20 +13,20 @@ jest.mock("react-i18next", () => ({
   useTranslation: () => ({ t: (key) => key }),
 }));
 
-describe("WaiverBlock", () => {
+describe("InDemandBlock", () => {
   const mockProps = {
     title: "test title",
     backgroundColorClass: "test-class",
   };
 
   it("renders title correctly", () => {
-    render(<WaiverBlock {...mockProps} />);
+    render(<InDemandBlock {...mockProps} />);
     expect(screen.getByText(mockProps.title)).toBeInTheDocument();
   });
 
   it("renders link correctly on tablet and larger devices", () => {
     (useMediaQuery as jest.Mock).mockReturnValue(true);
-    render(<WaiverBlock {...mockProps} />);
+    render(<InDemandBlock {...mockProps} />);
     const link = screen.getByText("OccupationPage.localAndRegionalWaiversText");
     expect(link).toBeInTheDocument();
     expect(link).toHaveAttribute(
@@ -37,12 +37,12 @@ describe("WaiverBlock", () => {
 
   it("does not render link on smaller devices", () => {
     (useMediaQuery as jest.Mock).mockReturnValue(false);
-    render(<WaiverBlock {...mockProps} />);
+    render(<InDemandBlock {...mockProps} />);
     expect(screen.queryByText("OccupationPage.localAndRegionalWaiversText")).toBeNull();
   });
 
   it("renders with correct class name", () => {
-    render(<WaiverBlock {...mockProps} />);
+    render(<InDemandBlock {...mockProps} />);
     expect(screen.getByText(mockProps.title).parentElement).toHaveClass(
       mockProps.backgroundColorClass,
     );
