+++ frontend/src/locales/en.ts
@@ -597,6 +597,11 @@ export const en = {
     providerApplication: "Icon showing folder with documents",
     careerOneStopLogo: "Official logo for Career One Stop organization",
   },
+  InDemandBlock: {
+    inDemandTitle: "In-Demand in all of New Jersey.",
+    localInDemandTitle: "In-Demand in {{countiesList}}.",
+    localAndRegionalWaiversText: "Learn about Local and Regional Waivers",
+  },
   InDemandPage: {
     pageTitle: "In-Demand Occupations",
     header: "In-Demand Occupations",
@@ -618,8 +623,6 @@ export const en = {
   OccupationPage: {
     header: "Occupation",
     dataUnavailableText: "This data is not yet available for this occupation.",
-    localExceptionCountiesTitle: "In-Demand in {{counties}}.",
-    localAndRegionalWaiversText: "Learn about Local and Regional Waivers",
     jobsOpenTitle: "Jobs Open in NJ",
     jobsOpenTooltip: "The number of jobs currently posted for this occupation in the State of NJ.",
     jobsOpenSource: "National Labor Exchange",
@@ -739,8 +742,6 @@ export const en = {
   },
   TrainingPage: {
     header: "Training Opportunity",
-    localExceptionCountiesTitle: "In-Demand in {{counties}}.",
-    localAndRegionalWaiversText: "Learn about Local and Regional Waivers",
     avgSalaryTitle: "Avg Salary after Program",
     avgSalaryTooltip:
       "Average salary 6 months after completion of this class or classes like it at this provider",
