+++ frontend/src/locales/es.ts
@@ -611,6 +611,11 @@ export const es = {
     providerApplication: "Icono que muestra la carpeta con documentos",
     careerOneStopLogo: "Logotipo oficial de la organización Career One Stop",
   },
+  InDemandBlock: {
+    inDemandTitle: "En demanda en todo Nueva Jersey.",
+    localExceptionCountiesTitle: "En demanda en {{counties}}.",
+    localAndRegionalWaiversText: "Infórmese sobre las exenciones locales y regionales",
+  },
   InDemandPage: {
     pageTitle: "Ocupaciones en demanda",
     header: "Ocupaciones en demanda",
@@ -630,8 +635,6 @@ export const es = {
   },
   OccupationPage: {
     header: "Ocupación",
-    localExceptionCountiesTitle: "En demanda en {{counties}}.",
-    localAndRegionalWaiversText: "Infórmese sobre las exenciones locales y regionales",
     dataUnavailableText: "Estos datos aún no están disponibles para esta ocupación.",
     jobsOpenTitle: "Trabajos abiertos en NJ",
     jobsOpenTooltip:
@@ -754,7 +757,7 @@ export const es = {
   },
   TrainingPage: {
     costGroupHeader: "Costo",
-    localExceptionCountiesTitle: "En demanda en {{counties}}.",
+    localExceptionCountiesTitle: "En demanda en {{countiesList}}.",
     localAndRegionalWaiversText: "Infórmese sobre las exenciones locales y regionales",
     header: "Oportunidad de entrenamiento",
     avgSalaryTitle: "Salario promedio después del programa",
